backend/requirements.txt
flask
flask-cors
python-dotenv
requests
python-jose[cryptography]

